Helicopter Ribbon and Ping Refactor
Refactoring how Operator's operations are actioned
Created Helicopter ribbon's visuals.  Needs a sound emitter and a
network message (like the sirens) to be complete.
Added InitEnableState so I can visualize operator and not have to worry
about remembering to re-enable it
A little positioning for the op's map to have it use less absolute
positioning